Implemented lazy loading for react-pdf and added export skeleton
